Refactor HDFS URI validation logic to correctly fix #25324
The fix for the issue just added a check for host 'default' that made the underlying hdfs library to treat it specially, searching for the HDFS connection string in $HADOOP_CONF_DIR/core-site.xml. However, using `from_uri` requires the URI to start with hdfs://, so the underlying library never receives exactly 'default'. With the change, both direct creation and `from_uri` works with an URI hdfs://default:0/...